+++ test/unit/request.test.js
@@ -10,6 +10,55 @@ describe('Request', function () {
     var rawRequest = fixtures.collectionV2.item[1].request,
         request = new Request(rawRequest);
 
+    describe('constructor', function () {
+        it('should handle all properties', function () {
+            var requestDefinition = {
+                    method: 'GET',
+                    url: {
+                        host: ['postman-echo', 'com'],
+                        'protocol': 'http',
+                        'query': [],
+                        'variable': []
+                    },
+                    auth: {
+                        type: 'basic',
+                        basic: [{
+                            key: 'username',
+                            type: 'string',
+                            value: 'postman'
+                        }, {
+                            key: 'password',
+                            type: 'string',
+                            value: 'password'
+                        }]
+                    }
+                },
+                request = new Request(requestDefinition);
+
+            expect(request).to.have.property('url');
+            expect(request.url).to.eql(new sdk.Url(requestDefinition.url));
+            expect(request).to.have.property('auth');
+            expect(request.auth).to.eql(new sdk.RequestAuth(requestDefinition.auth));
+        });
+
+        it('should not create auth if auth is falsy', function () {
+            var requestDefinition = {
+                    method: 'GET',
+                    url: {
+                        host: ['postman-echo', 'com'],
+                        'protocol': 'http',
+                        'query': [],
+                        'variable': []
+                    },
+                    auth: null
+                },
+                request = new Request(requestDefinition);
+
+            expect(request).to.not.have.property('auth');
+            expect(request.toJSON()).to.not.have.property('auth');
+        });
+    });
+
     describe('sanity', function () {
         var rawRequest = fixtures.collectionV2.item[1],
             request = new Request(rawRequest.request);
@@ -150,7 +199,13 @@ describe('Request', function () {
             var jsonified = request.toJSON();
 
             expect(jsonified.method).to.eql(rawRequest.method);
-            expect(jsonified.url).to.eql(rawRequest.url);
+            expect(jsonified.url).to.eql({
+                protocol: 'http',
+                path: ['post'],
+                host: ['echo', 'getpostman', 'com'],
+                query: [],
+                variable: []
+            });
             expect(jsonified.header).to.eql(rawRequest.header);
             expect(jsonified.body).to.eql(rawRequest.body);
             expect(jsonified.description).to.eql(rawRequest.description);